introduce ParsedGrammar, make PEGTransformerFactory entirely static
