+++ src/parser/peg/CMakeLists.txt
@@ -13,6 +13,7 @@ add_library_unity(
   matcher_factory.cpp
   parser_packrat.cpp
   peg_parser.cpp
+  parsed_grammar.cpp
   sql_formatter.cpp)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_parser_peg>
