+++ src/parser/peg/compiled_grammar.cpp
@@ -3,11 +3,6 @@
 #include "duckdb/parser/peg/keyword_helper/duckdb_keyword_helper.hpp"
 #include "duckdb/main/database.hpp"
 #include "duckdb/main/client_config.hpp"
-#ifdef PEG_PARSER_SOURCE_FILE
-#include <fstream>
-#else
-#include "duckdb/parser/peg/inlined_grammar.hpp"
-#endif
 
 namespace duckdb {
 
@@ -34,11 +29,16 @@ shared_ptr<CompiledGrammar> CompiledGrammar::Get(DatabaseInstance &db) {
 	return parser_cache.GetMatcher(nullptr);
 }
 
-const PEGTransformerFactory &CompiledGrammar::GetTransformerFactory() {
-	return transformer_factory;
+ParserCache::ParserCache() : version(0) {
 }
 
-ParserCache::ParserCache() : version(0) {
+static void ValidateParsedGrammarRoots(const ParsedGrammar &grammar) {
+	if (!grammar.HasRule("Program")) {
+		throw InvalidInputException("Grammar is missing required root rule 'Program'");
+	}
+	if (!grammar.HasRule("TopLevelStatement")) {
+		throw InvalidInputException("Grammar is missing required root rule 'TopLevelStatement'");
+	}
 }
 
 shared_ptr<CompiledGrammar> ParserCache::GetMatcher(optional_ptr<ClientContext> context) {
@@ -48,20 +48,39 @@ shared_ptr<CompiledGrammar> ParserCache::GetMatcher(optional_ptr<ClientContext> 
 			return matcher;
 		}
 	}
-	auto new_matcher = shared_ptr<CompiledGrammar>(new CompiledGrammar(*this));
-	MatcherFactory factory(new_matcher->allocator, parsed_grammar, *new_matcher);
-#ifdef PEG_PARSER_SOURCE_FILE
-	std::ifstream t(PEG_PARSER_SOURCE_FILE);
-	std::stringstream buffer;
-	buffer << t.rdbuf();
-	auto grammar_string = buffer.str();
 
-	new_matcher->program_matcher = factory.CreateMatcher(grammar_string.c_str(), "Program");
-#else
-	new_matcher->program_matcher = factory.CreateMatcher(const_char_ptr_cast(INLINED_PEG_GRAMMAR), "Program");
-#endif
-	// TopLevelStatement is referenced by Program, so it has already been built and cached.
+	auto grammar = ParsedGrammar::CreateDefault();
+	ValidateParsedGrammarRoots(grammar);
+	for (auto &entry : grammar.rules) {
+		auto &rule = *entry.second;
+		if (rule.semantic && !rule.transform) {
+			throw InvalidInputException("Semantic grammar rule '%s' has no transform function", rule.name);
+		}
+		for (auto &token : rule.recipe.tokens) {
+			if (token.type != PEGTokenType::REFERENCE && token.type != PEGTokenType::FUNCTION_CALL) {
+				continue;
+			}
+			if (token.type == PEGTokenType::REFERENCE && rule.recipe.parameters.count(token.text)) {
+				continue;
+			}
+			if (!grammar.HasRule(token.text.GetString())) {
+				throw InvalidInputException("Grammar rule '%s' references missing rule '%s'", rule.name,
+				                            token.text.GetString());
+			}
+		}
+	}
+
+	auto new_matcher = shared_ptr<CompiledGrammar>(new CompiledGrammar(*this));
+	for (auto &entry : grammar.rules) {
+		auto &rule = *entry.second;
+		new_matcher->rules.emplace(rule.name, make_uniq<CompiledGrammarRule>(rule.name, std::move(rule.transform),
+		                                                                     std::move(rule.trampoline_transform),
+		                                                                     std::move(rule.trampoline_ops)));
+	}
+	MatcherFactory factory(new_matcher->allocator, grammar, *new_matcher);
+	new_matcher->program_matcher = factory.CreateRootMatcher("Program");
 	new_matcher->top_level_statement_matcher = factory.GetMatcher("TopLevelStatement");
+
 	std::unique_lock<std::mutex> lock(mutex);
 	if (!matcher) {
 		matcher = std::move(new_matcher);
