+++ src/parser/peg/peg_parser.cpp
@@ -202,124 +202,4 @@ void PEGParser::ParseRules(const char *grammar) {
 	}
 }
 
-ParsedGrammar::ParsedGrammar(ParsedGrammar &&other) : rules(std::move(other.rules)) {
-	string_heap.Move(other.string_heap);
-}
-
-ParsedGrammar &ParsedGrammar::operator=(ParsedGrammar &&other) {
-	if (this != &other) {
-		rules.clear();
-		string_heap.Destroy();
-		string_heap.Move(other.string_heap);
-		rules = std::move(other.rules);
-	}
-	return *this;
-}
-
-ParsedGrammar ParsedGrammar::Parse(const string &grammar) {
-	PEGParser parser;
-	parser.ParseRules(grammar.c_str());
-	ParsedGrammar result;
-	for (auto &entry : parser.rules) {
-		result.AddParsedRule(ParsedGrammarRule(entry.first, std::move(entry.second)));
-	}
-	return result;
-}
-
-ParsedGrammar ParsedGrammar::CreateDefault() {
-	auto result = Parse(INLINED_PEG_GRAMMAR);
-	if (!result.HasRule("EndOfInput")) {
-		result.AddParsedRule(ParsedGrammarRule("EndOfInput", PEGRule()));
-	}
-	PEGTransformerFactory::RegisterDefaultTransforms(result);
-	return result;
-}
-
-bool ParsedGrammar::HasRule(const string &rule_name) const {
-	return rules.find(rule_name) != rules.end();
-}
-
-const ParsedGrammarRule &ParsedGrammar::GetRule(const string &rule_name) const {
-	auto entry = rules.find(rule_name);
-	if (entry == rules.end()) {
-		throw InvalidInputException("Grammar rule '%s' does not exist", rule_name);
-	}
-	return *entry->second;
-}
-
-ParsedGrammarRule &ParsedGrammar::GetMutableRule(const string &rule_name) {
-	return const_cast<ParsedGrammarRule &>(GetRule(rule_name));
-}
-
-ParsedGrammarRule ParsedGrammar::ParseSingleRule(const string &rule_definition) {
-	PEGParser parser;
-	parser.ParseRules(rule_definition.c_str());
-	if (parser.rules.size() != 1) {
-		throw InvalidInputException("Expected exactly one PEG rule definition");
-	}
-	auto &entry = *parser.rules.begin();
-	return ParsedGrammarRule(entry.first, std::move(entry.second));
-}
-
-void ParsedGrammar::RegisterStrings(PEGRule &rule) {
-	string_map_t<idx_t> parameters;
-	for (auto &entry : rule.parameters) {
-		parameters.emplace(string_heap.AddString(entry.first), entry.second);
-	}
-	rule.parameters = std::move(parameters);
-	for (auto &token : rule.tokens) {
-		token.text = string_heap.AddString(token.text);
-	}
-}
-
-void ParsedGrammar::AddParsedRule(ParsedGrammarRule rule) {
-	if (HasRule(rule.name)) {
-		throw InvalidInputException("Grammar rule '%s' already exists", rule.name);
-	}
-	RegisterStrings(rule.recipe);
-	auto name = rule.name;
-	rules.emplace(std::move(name), make_uniq<ParsedGrammarRule>(std::move(rule)));
-}
-
-void ParsedGrammar::AddRule(const string &rule_definition, grammar_transform_function_t transform,
-                            grammar_transform_function_t trampoline_transform) {
-	auto rule = ParseSingleRule(rule_definition);
-	rule.transform = std::move(transform);
-	rule.trampoline_transform = trampoline_transform ? std::move(trampoline_transform) : rule.transform;
-	rule.semantic = bool(rule.transform);
-	AddParsedRule(std::move(rule));
-}
-
-void ParsedGrammar::ReplaceRule(const string &rule_definition, grammar_transform_function_t transform,
-                                grammar_transform_function_t trampoline_transform) {
-	auto rule = ParseSingleRule(rule_definition);
-	auto entry = rules.find(rule.name);
-	if (entry == rules.end()) {
-		throw InvalidInputException("Grammar rule '%s' does not exist", rule.name);
-	}
-	RegisterStrings(rule.recipe);
-	rule.transform = std::move(transform);
-	rule.trampoline_transform = trampoline_transform ? std::move(trampoline_transform) : rule.transform;
-	rule.semantic = bool(rule.transform);
-	entry->second = make_uniq<ParsedGrammarRule>(std::move(rule));
-}
-
-void ParsedGrammar::RemoveRule(const string &rule_name) {
-	if (rules.erase(rule_name) == 0) {
-		throw InvalidInputException("Grammar rule '%s' does not exist", rule_name);
-	}
-}
-
-void ParsedGrammar::SetTransform(const string &rule_name, grammar_transform_function_t transform,
-                                 grammar_transform_function_t trampoline_transform) {
-	auto &rule = GetMutableRule(rule_name);
-	rule.transform = std::move(transform);
-	rule.trampoline_transform = trampoline_transform ? std::move(trampoline_transform) : rule.transform;
-	rule.semantic = true;
-}
-
-void ParsedGrammar::SetTrampolineOps(const string &rule_name, const TransformFrameOps &ops) {
-	GetMutableRule(rule_name).trampoline_ops = make_shared_ptr<TransformFrameOps>(ops);
-}
-
 } // namespace duckdb
